refactor: common 폴더의 가독성을 위해  IdolImage 폴더 추가
