+++ src/components/common/IdolImage/IdolImage.jsx
