+++ src/components/common/IdolImage/IdolImage.scss
@@ -1,4 +1,4 @@
-@import "../../styles/color.scss";
+@import "../../../styles/color.scss";
 
 .imgBorder {
   border-radius: 50%;
